+++ docs/nuxt.config.ts
@@ -194,7 +194,8 @@ const pages = [
   '/docs/composables/use-confetti/',
   '/docs/composables/use-overlay/',
   '/docs/composables/use-toast/',
-  '/docs/composables/use-speech-recognition/'
+  '/docs/composables/use-speech-recognition/',
+  '/docs/composables/use-device/'
   // endregion ////
 ]