+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## [2.5.2](https://github.com/bitrix24/b24ui/compare/v2.5.1...v2.5.2) (2026-03-xx)
 
+* feat(useDevice): new composables for detect the current platform (Bitrix24 mobile/desktop app or web) and screen size
 * fix(NavigationMenu): improve theme
 * fix(DashboardSidebar|Header): improve menu
 * docs(dashboard*): improve